@@ -7,17 +7,10 @@
   <packaging>jar</packaging> 
 
   <dependencies>
-       		<dependency>
-            		<groupId>javax.servlet</groupId>
-            		<artifactId>javax.servlet-api</artifactId>
-            		<version>4.0.1</version>
-            		<scope>provided</scope>
-        	</dependency>
-
         <dependency>
 		    <groupId>fr.cirad</groupId>
 		    <artifactId>mgdb2</artifactId>
-		    <version>2.5.2-alpha</version>
+		    <version>2.6-beta</version>
             <exclusions>
                 <exclusion>
                     <groupId>com.fasterxml.jackson.core</groupId>
@@ -35,7 +28,7 @@
 		<dependency>
 			<groupId>fr.cirad</groupId>
 			<artifactId>role_manager</artifactId>
-			<version>1.3.2-RELEASE</version>
+			<version>1.3.3-RELEASE</version>
 		</dependency>
 
         <dependency>
@@ -53,7 +46,7 @@
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-webmvc</artifactId>
-            <version>5.3.23</version>
+            <version>5.3.25</version>
             <exclusions>
                 <exclusion>
                     <groupId>commons-logging</groupId>
@@ -96,12 +89,12 @@
 		</dependency>
                 
         <!-- https://mvnrepository.com/artifact/javax.servlet/javax.servlet-api -->
-        <dependency>
-            <groupId>javax.servlet</groupId>
-            <artifactId>javax.servlet-api</artifactId>
-            <version>4.0.1</version>
-            <scope>provided</scope>
-        </dependency>
+<!--        <dependency>-->
+<!--            <groupId>javax.servlet</groupId>-->
+<!--            <artifactId>javax.servlet-api</artifactId>-->
+<!--            <version>4.0.1</version>-->
+<!--            <scope>provided</scope>-->
+<!--        </dependency>-->
   	</dependencies>
   <repositories>
 	  <repository>
@@ -117,8 +110,8 @@
         <artifactId>maven-compiler-plugin</artifactId>
         <version>3.5.1</version>
         <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
+          <source>17</source>
+          <target>17</target>
         </configuration>
       </plugin>
 <!--       <plugin> -->
